handleres for node was moved to `api_node.go`; thanks @Charleslee522
